addとeditの共通処理をkindness_giver_base_view_modelにまとめる
